refactored to match calculator example
