+++ kamodo/kamodo.py
@@ -5,7 +5,7 @@ No Copyright is claimed in the United States under Title 17, U.S. Code.  All Oth
 """
 import asyncio
 import os
-import ssl
+
 
 from util import np
 from sympy import Integral, Symbol, symbols, Function
@@ -78,6 +78,7 @@ import sympy
 
 from rpc.proto import capnp, KamodoRPC, FunctionRPC, kamodo_capnp, rpc_map_to_dict
 from rpc.proto import from_rpc_literal, to_rpc_literal, to_rpc_expr
+from rpc.proto import Server
 import socket
 from util import construct_signature
 
@@ -963,7 +964,7 @@ class Kamodo(UserDict):
             func=FunctionRPC(func),
             meta=self.to_rpc_meta(key),
         )
-        self._server[key] = field
+        self._kamodo_rpc[key] = field
 
     # def serve(self, socket_='*:60000'):
     #     self._server = KamodoRPC()
@@ -981,115 +982,16 @@ class Kamodo(UserDict):
     #         server = capnp.TwoPartyServer(socket_, bootstrap=self._server)
     #         server.run_forever()
 
-    class Server():
-        async def server_reader(self):
-            """
-            Reader for the server side.
-            """
-            while self.retry:
-                try:
-                    # Must be a wait_for so we don't block on read()
-                    data = await asyncio.wait_for(
-                        self.reader.read(4096),
-                        timeout=0.1
-                    )
-                except asyncio.TimeoutError:
-                    print("reader timeout.")
-                    continue
-                except Exception as err:
-                    print("Unknown reader err: %s", err)
-                    return False
-                await self.server.write(data)
-            print("reader done.")
-            return True
-
-        async def server_writer(self):
-            """
-            Writer for the server side.
-            """
-            while self.retry:
-                try:
-                    # Must be a wait_for so we don't block on read()
-                    data = await asyncio.wait_for(
-                        self.server.read(4096),
-                        timeout=0.1
-                    )
-                    self.writer.write(data.tobytes())
-                except asyncio.TimeoutError:
-                    print("writer timeout.")
-                    continue
-                except Exception as err:
-                    print("Unknown writer err: %s", err)
-                    return False
-            print("writer done.")
-            return True
-
-        async def kamodo_server(self, reader, writer):
-            # Start TwoPartyServer using TwoWayPipe (only requires bootstrap)
-            self.server = capnp.TwoPartyServer(bootstrap=KamodoRPC())
-            self.reader = reader
-            self.writer = writer
-            self.retry = True
-
-            # Assemble reader and writer tasks, run in the background
-            coroutines = [self.server_reader(), self.server_writer()]
-            tasks = asyncio.gather(*coroutines, return_exceptions=True)
-
-            while True:
-                self.server.poll_once()
-                # Check to see if reader has been sent an eof (disconnect)
-                if self.reader.at_eof():
-                    self.retry = False
-                    break
-                await asyncio.sleep(0.01)
-
-            # Make wait for reader/writer to finish (prevent possible resource leaks)
-            await tasks
-
-    async def new_connection(self, reader, writer):
-        server = self.Server()
-        await server.kamodo_server(reader, writer)
-
-    async def serve(self):
-
-        """
-        Method to start communication as asynchronous server.
-        """
-        addr = 'localhost'
-        port = '60000'
-
-        self._server = KamodoRPC()
+    def serve(self):
+        # Register rpc fields
+        self._kamodo_rpc = KamodoRPC()
         for key in self.signatures:
             if self.verbose:
                 print('serving {}'.format(key))
             self.register_rpc_field(key)
 
-        ctx = ssl.create_default_context(ssl.Purpose.CLIENT_AUTH)
-        this_dir = os.path.dirname(os.path.abspath(__file__))
-        print(f"THIS DIR : {this_dir}")
-        ctx.load_cert_chain(
-            os.path.join(this_dir, "selfsigned.cert"),
-            os.path.join(this_dir, "selfsigned.key"),
-        )
-
-        # Handle both IPv4 and IPv6 cases
-        try:
-            print("Try IPv4")
-            server = await asyncio.start_server(
-                self.new_connection,
-                addr, port, ssl=ctx,
-                family=socket.AF_INET
-            )
-        except Exception:
-            print("Try IPv6")
-            server = await asyncio.start_server(
-                self.new_connection,
-                addr, port, ssl=ctx,
-                family=socket.AF_INET6
-            )
-
-        async with server:
-            await server.serve_forever()
+        self.async_server = Server(self._kamodo_rpc)
+        asyncio.run(self.async_server.serve())
 
     def figure(self, variable, indexing='ij', **kwargs):
         """Generates a plotly figure for a given variable and keyword arguments"""
